requirements.txt
annotated-types==0.6.0
APScheduler==3.10.4
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
black==24.4.2
click==8.1.7
colorama==0.4.6
crontab==1.0.1
decorator==5.1.1
distribution==1.0.0
Django==4.2.2
django-apscheduler==0.6.2
django-bootstrap-datepicker-plus==5.0.5
django-cron==0.6.0
django-crontab==0.7.1
django-extensions==3.2.3
django-phonenumber-field==7.3.0
django-widget-tweaks==1.5.0
executing==2.0.1
install==1.3.5
ipython==8.24.0
jedi==0.19.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
phonenumbers==8.13.38
pillow==10.3.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.4
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
wcwidth==0.2.13